PS-2133 Display Program-Specific Master Data

+++ src/adapters/postgres/fields-adapter.ts
@@ -191,7 +191,7 @@ export class PostgresFieldsService implements IServicelocatorfields {
         return schema;
     }
 
-    async createFields(request: any, fieldsDto: FieldsDto, response: Response,) {
+    async createFields(request: any, fieldsDto: FieldsDto, tenantId: String, response: Response,) {
         const apiId = APIID.FIELDS_CREATE;
         try {
             const fieldsData: any = {}; // Define an empty object to store field data
@@ -219,14 +219,9 @@ export class PostgresFieldsService implements IServicelocatorfields {
                     "name": fieldsData.name
                 }
             })
+
             if (checkFieldExist.length > 0) {
-                APIResponse.error(
-                    response,
-                    apiId,
-                    `Fields already exist`,
-                    `CONFLICT`,
-                    (HttpStatus.CONFLICT)
-                )
+                APIResponse.error(response, apiId, `Fields already exist`, `CONFLICT`, (HttpStatus.CONFLICT))
             }
 
             let storeWithoutControllingField = [];
@@ -241,10 +236,16 @@ export class PostgresFieldsService implements IServicelocatorfields {
                     let query = `SELECT COUNT(*) FROM public.${fieldsData.sourceDetails.table} WHERE value = '${sourceFieldName['value']}'`;
                     const checkSourceData = await this.fieldsValuesRepository.query(query);
 
+                    if (!tenantId) {
+                        return await APIResponse.error(
+                            response, apiId, 'Tenant ID is required', 'TENANT_ID_MISSING', HttpStatus.BAD_REQUEST
+                        );
+                    }
+
                     if (checkSourceData[0].count == 0) {
-                        let createSourceField = await this.createSourceDetailsTableFields(fieldsData.sourceDetails.table, sourceFieldName['name'], sourceFieldName['value'], createdBy, sourceFieldName['controllingfieldfk'], fieldsData?.dependsOn);
+                        let createSourceField = await this.createSourceDetailsTableFields(fieldsData.sourceDetails.table, sourceFieldName['name'], sourceFieldName['value'], createdBy, tenantId, sourceFieldName['controllingfieldfk'], fieldsData?.dependsOn);
                     } else {
-                        let updateSourceField = await this.updateSourceDetailsTableFields(fieldsData.sourceDetails.table, sourceFieldName['name'], sourceFieldName['value'], updatedBy, sourceFieldName['controllingfieldfk']);
+                        let updateSourceField = await this.updateSourceDetailsTableFields(fieldsData.sourceDetails.table, sourceFieldName['name'], sourceFieldName['value'], tenantId, updatedBy, sourceFieldName['controllingfieldfk']);
                     }
                 }
                 delete fieldsData.fieldParams;
@@ -266,7 +267,7 @@ export class PostgresFieldsService implements IServicelocatorfields {
         }
     }
 
-    async updateFields(fieldId: any, request: any, fieldsUpdateDto: FieldsUpdateDto, response: Response) {
+    async updateFields(fieldId: any, request: any, fieldsUpdateDto: FieldsUpdateDto, tenantId: String, response: Response) {
         const apiId = APIID.FIELDS_CREATE;
         try {
             const decoded: any = jwt_decode(request.headers.authorization);
@@ -304,15 +305,26 @@ export class PostgresFieldsService implements IServicelocatorfields {
                         storeWithoutControllingField.push(sourceFieldName['name'])
                     }
 
-                    // check options exits in source table column or not
-                    let query = `SELECT COUNT(*) FROM public.${getSourceDetails.sourceDetails.table} WHERE value = '${sourceFieldName['value']}'`;
+                    if (!tenantId) {
+                        return await APIResponse.error(
+                            response, apiId, 'Tenant ID is required', 'TENANT_ID_MISSING', HttpStatus.BAD_REQUEST
+                        );
+                    }
+
+                    // check options exits in source table column or not for that particular tenant
+                    let query = `SELECT COUNT(*) FROM public.${getSourceDetails.sourceDetails.table} WHERE "value" = '${sourceFieldName['value']}' AND "tenantId" = '${tenantId}'`;
+
                     const checkSourceData = await this.fieldsValuesRepository.query(query);
+
+
+
+
 
                     //If not exist then create that column else update that data
                     if (checkSourceData[0].count == 0) {
-                        let createSourceField = await this.createSourceDetailsTableFields(getSourceDetails.sourceDetails.table, sourceFieldName['name'], sourceFieldName['value'], createdBy, sourceFieldName['controllingfieldfk'], getSourceDetails.dependsOn);
+                        let createSourceField = await this.createSourceDetailsTableFields(getSourceDetails.sourceDetails.table, sourceFieldName['name'], sourceFieldName['value'], createdBy, tenantId, sourceFieldName['controllingfieldfk'], getSourceDetails.dependsOn);
                     } else {
-                        let updateSourceField = await this.updateSourceDetailsTableFields(getSourceDetails.sourceDetails.table, sourceFieldName['name'], sourceFieldName['value'], updatedBy, sourceFieldName['controllingfieldfk']);
+                        let updateSourceField = await this.updateSourceDetailsTableFields(getSourceDetails.sourceDetails.table, sourceFieldName['name'], sourceFieldName['value'], tenantId, updatedBy, sourceFieldName['controllingfieldfk']);
                     }
                 }
                 delete fieldsData.fieldParams;
@@ -384,16 +396,16 @@ export class PostgresFieldsService implements IServicelocatorfields {
 
     }
 
-    async createSourceDetailsTableFields(tableName: string, name: string, value: string, createdBy: string, controllingfieldfk?: string, dependsOn?: string) {
+    async createSourceDetailsTableFields(tableName: string, name: string, value: string, createdBy: string, tenantId: String, controllingfieldfk?: string, dependsOn?: string) {
 
-        let createSourceFields = `INSERT INTO public.${tableName} ("name", "value", "createdBy"`;
+        let createSourceFields = `INSERT INTO public.${tableName} ("name", "value", "createdBy","tenantId"`;
 
         // Add controllingfieldfk to the columns if it is defined
         if (controllingfieldfk !== undefined && controllingfieldfk !== '') {
             createSourceFields += `, controllingfieldfk`;
         }
 
-        createSourceFields += `) VALUES ('${name}', '${value}', '${createdBy}'`;
+        createSourceFields += `) VALUES ('${name}', '${value}', '${createdBy}', '${tenantId}'`;
 
         // Add controllingfieldfk to the values if it is defined
         if (controllingfieldfk !== undefined && controllingfieldfk !== '') {
@@ -413,7 +425,7 @@ export class PostgresFieldsService implements IServicelocatorfields {
         }
     }
 
-    async updateSourceDetailsTableFields(tableName: string, name: string, value: string, updatedBy: string, controllingfieldfk?: string) {
+    async updateSourceDetailsTableFields(tableName: string, name: string, value: string, tenantId: String, updatedBy: string, controllingfieldfk?: string) {
 
         let updateSourceDetails = `UPDATE public.${tableName} SET "name"='${name}',"updatedBy"='${updatedBy}'`;
 
@@ -421,7 +433,7 @@ export class PostgresFieldsService implements IServicelocatorfields {
             updateSourceDetails += `, controllingfieldfk='${controllingfieldfk}'`;
         }
 
-        updateSourceDetails += ` WHERE value='${value}';`;
+        updateSourceDetails += ` WHERE value='${value}' AND "tenantId" = '${tenantId}';`;
 
         const updateSourceData = await this.fieldsValuesRepository.query(updateSourceDetails);
         if (updateSourceData.length == 0) {
@@ -760,7 +772,7 @@ export class PostgresFieldsService implements IServicelocatorfields {
     }
 
     //Get all fields options
-    public async getFieldOptions(fieldsOptionsSearchDto: FieldsOptionsSearchDto, response: Response) {
+    public async getFieldOptions(fieldsOptionsSearchDto: FieldsOptionsSearchDto, tenantId: string, response: Response) {
         const apiId = APIID.FIELDVALUES_SEARCH;
         try {
 
@@ -795,12 +807,21 @@ export class PostgresFieldsService implements IServicelocatorfields {
             }
 
             if (fetchFieldParams?.sourceDetails?.source === 'table') {
+                //If field option source is a table then tenantId is required
+
+                if (!tenantId) {
+                    return await APIResponse.error(
+                        response, apiId, 'Tenant ID is required', 'TENANT_ID_MISSING', HttpStatus.BAD_REQUEST
+                    );
+                }
+
                 let whereClause;
                 if (controllingfieldfk) {
                     whereClause = `"controllingfieldfk" = '${controllingfieldfk}'`;
                 }
 
-                dynamicOptions = await this.findDynamicOptions(fieldName, whereClause, offset, limit, order, optionName);
+                dynamicOptions = await this.findDynamicOptions(fieldName, whereClause, offset, limit, order, optionName, tenantId);
+
             } else if (fetchFieldParams?.sourceDetails?.source === 'jsonFile') {
                 const filePath = path.join(
                     process.cwd(),
@@ -829,6 +850,7 @@ export class PostgresFieldsService implements IServicelocatorfields {
             const queryData = dynamicOptions.map(result => ({
                 value: result?.value,
                 label: result?.name,
+                tenantId: result?.tenantId,
                 createdAt: result?.createdAt,
                 updatedAt: result?.updatedAt,
                 createdBy: result?.createdBy,
@@ -886,8 +908,22 @@ export class PostgresFieldsService implements IServicelocatorfields {
 
             //Delete data from source table
             if (getField?.sourceDetails?.source == 'table') {
+
+                if (!requiredData.tenantId) {
+                    return await APIResponse.error(
+                        response, apiId, 'Tenant ID is required', 'TENANT_ID_MISSING', HttpStatus.BAD_REQUEST
+                    );
+                }
+
                 let whereCond = requiredData.option ? `WHERE "value"='${requiredData.option}'` : '';
+                if (whereCond) {
+                    whereCond += `AND "tenantId"='${requiredData.tenantId}'`
+                } else {
+                    whereCond = `WHERE "tenantId"='${requiredData.tenantId}'`
+                }
+
                 let query = `DELETE FROM public.${getField?.sourceDetails?.table} ${whereCond}`
+
                 let [_, affectedRow] = await this.fieldsRepository.query(query);
 
                 if (affectedRow === 0) {
@@ -927,7 +963,7 @@ export class PostgresFieldsService implements IServicelocatorfields {
         }
     }
 
-    async findDynamicOptions(tableName, whereClause?: {}, offset?: {}, limit?: {}, order?: {}, optionName?: {}) {
+    async findDynamicOptions(tableName, whereClause?: {}, offset?: {}, limit?: {}, order?: {}, optionName?: {}, tenantId?: string) {
         let query: string;
         let result;
 
@@ -946,6 +982,11 @@ export class PostgresFieldsService implements IServicelocatorfields {
         } else {
             whereCond += ''
         }
+
+        if (tenantId) {
+            whereCond += whereCond ? ` AND "tenantId" = '${tenantId}'` : `WHERE "tenantId" = '${tenantId}'`;
+        }
+
 
         query = `SELECT *,COUNT(*) OVER() AS total_count FROM public."${tableName}" ${whereCond} ${orderCond} ${offsetCond} ${limitCond}`
